Fix cyclesim CSR_PDS workgroup semantics
Treat CSR_PDS as the resident workgroup PDS slot base instead of a per-warp base. Subcore dispatch now writes the same workgroup slot base for every warp in the workgroup.

Carry CSR_TID, CSR_NUMW, and CSR_NUMT into LSU requests so private vector load/store instructions can compute the Spike/GVM-compatible workgroup-interleaved PDS address: CSR_PDS + align4(offset) * numw * numt + (tid + lane) * 4, preserving byte-lane offset only for byte/half private accesses.

Size and allocate the cyclesim PDS virtual memory as the resident slot pool rather than as a full grid-wide private buffer. This matches the resident-slot contract used by dispatch and avoids baking global workgroup ids into PDS addressing.

Update the metadata documentation to describe pdsBaseAddr as the resident private-memory pool base and pdsSize as per-thread private memory size.

Verified with the OpenCL pds_absolute_alias_probe across cycle, rtl-nocache, and rtl-with-cache backends using /work/ventus-env-github/install; all reported mismatches=0.

+++ src/task.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include "ventus_cyclesim.h"
 #include <any>
+#include <cstddef>
 #include <functional>
 #include <map>
 #include <memory>
@@ -18,7 +19,7 @@ public:
     task_t(
         uint32_t id, const std::string name, uint64_t pagetable,
         std::function<void()> finish_callback = nullptr,
-        std::function<void(uint32_t, uint32_t)> vmem_free = nullptr
+        std::function<void(uint32_t, size_t)> vmem_free = nullptr
     );
 
     void add_kernel(std::shared_ptr<kernel_t> kernel);
@@ -61,8 +62,8 @@ private:
     enum { TASKSTATUS_IDLE, TASKSTATUS_RUNNING, TASKSTATUS_FINISHED } m_status = TASKSTATUS_IDLE;
 
     uint32_t m_kernel_private_memory_vaddr = 0;
-    uint32_t m_kernel_private_memory_size = 0;
+    size_t m_kernel_private_memory_size = 0;
 
     std::function<void()> m_finish_callback = nullptr; // task finish callback
-    std::function<void(uint32_t vaddr, uint32_t size)> m_vmem_free = nullptr;
+    std::function<void(uint32_t vaddr, size_t size)> m_vmem_free = nullptr;
 };
